--- src/gnu/Binomial_random123.cpp
@@ -1,7 +1,0 @@
-#include <RNG_random123.h>
-#include <Binomial_random123.h>
-
-double Binomial_random123::operator()()
-{
-    return(d(*(generator()->get_random_generator())));
-};
