--- src/gnu/Geom_random123.cpp
@@ -1,7 +1,0 @@
-#include <RNG_random123.h>
-#include <Geom_random123.h>
-
-double Geometric_random123::operator()()
-{
-    return(1. + d(*(generator()->get_random_generator())));
-};
