--- src/gnu/NegExp_random123.cpp
@@ -1,7 +1,0 @@
-#include <RNG_random123.h>
-#include <NegExp_random123.h>
-
-double NegativeExpntl_random123::operator()()
-{
-    return(d(*(generator()->get_random_generator())));
-};
