--- src/gnu/Poisson_random123.cpp
@@ -1,7 +1,0 @@
-#include <RNG_random123.h>
-#include <Poisson_random123.h>
-
-double Poisson_random123::operator()()
-{
-    return(d(*(generator()->get_random_generator())));
-};
